+++ src/factory/action/check/paymentMethod/movieTicket.ts
@@ -6,20 +6,17 @@ import { IMovieTicket } from '../../../paymentMethod/paymentCard/movieTicket';
 import * as CheckActionFactory from '../../check';
 import * as PayActionFactory from '../../trade/pay';
 
-export type IAgent = ActionFactory.IParticipant;
+export type IAgent = ActionFactory.IParticipantAsWebApplication | ActionFactory.IParticipantAsPerson;
 export type IPurchaseNumberAuthIn = surfrock.service.auth.factory.IPurchaseNumberAuthIn;
 export type IPurchaseNumberAuthResult = surfrock.service.auth.factory.IPurchaseNumberAuthResult;
 export type IMovieTicketResult = IMovieTicket;
-
 export interface IPaymentService extends PayActionFactory.IPaymentService {
     /**
      * 販売者
      */
     seller: { typeOf: OrganizationType; id: string };
 }
-
 export type IObject = IPaymentService[];
-
 export interface IResult {
     purchaseNumberAuthIn: IPurchaseNumberAuthIn;
     purchaseNumberAuthResult: IPurchaseNumberAuthResult;
@@ -28,18 +25,15 @@ export interface IResult {
      */
     movieTickets: IMovieTicketResult[];
 }
-
 export type IError = any;
-
 /**
- * ムビチケ確認アクション属性インターフェース
+ * ムビチケ確認アクション属性
  */
 export interface IAttributes extends CheckActionFactory.IAttributes<IObject, IResult> {
-    object: IObject;
     agent: IAgent;
+    object: IObject;
 }
-
 /**
- * ムビチケ確認アクションインターフェース
+ * ムビチケ確認アクション
  */
 export type IAction = ActionFactory.IAction<IAttributes>;