+++ src/factory/action/transfer/give/pointAward.ts
@@ -1,15 +1,12 @@
+import * as ActionFactory from '../../../action';
 import * as OrderFactory from '../../../order';
-import { OrganizationType } from '../../../organizationType';
 import { ObjectType as AuthorizePointAwardActionObjectType } from '../../authorize/award/point';
 import * as GiveActionFactory from '../give';
 
-export interface IAgent {
-    id: string;
-    typeOf: OrganizationType.Project;
-}
+export type IAgent = ActionFactory.IParticipantAsProject;
 export import ObjectType = AuthorizePointAwardActionObjectType;
 /**
- * ポイント特典付与対象インターフェース
+ * ポイント特典付与対象
  */
 export interface IObject {
     typeOf: ObjectType;
@@ -36,21 +33,17 @@ export interface IObject {
     };
     description: string;
 }
-
 export type IResult = any;
-
 /**
  * 目的は注文
  * 注文に対する特典としてポイントが付与される仕組み
  */
 export type IPurpose = OrderFactory.ISimpleOrder;
-
 export type IPotentialActions = any;
 export interface IAttributes extends GiveActionFactory.IAttributes<IObject, IResult> {
     agent: IAgent;
     purpose: IPurpose;
 }
-
 /**
  * ポイント特典付与アクション
  */